fix(xv6-mbc): mmio_putc byte-store eliminates 2-null pad per char
Per the followup flagged in commit 3ac1f684 ("2-null-pad-per-char TTY
waste from unaligned word store"). The previous u32 store at unaligned
address 0xC001 was lowered by GCC into 4 separate byte stores at
0xC001-0xC004, of which the first 3 hit the eBPF MMIO TTY intercept
range and emitted char + 2 zero pads per call.

Changed UPC_TTY_REG to `volatile uint8 *` and the cast to `(uint8)c`.
The compiler now emits a single `sb` at 0xC001, the intercept fires
once, the operator sees the actual char.

Live boot result on kernel 6.17.0-23:
  Before: 45 bytes "x..v..6.. ..b..o..o..t..i..n..g..\\..\\..\\.\\..."
  After:  15 bytes "xv6 booting...\\n"  (clean ASCII)

66% TTY bandwidth reduction. Banner detection still fires the gate
banner. Cosmetic-only — no behavioural change to the kernel.

+++ crates/xv6-mbc/adapters/start_mbc.c
@@ -80,11 +80,22 @@ struct boot_params_v2 {
 #define MSTATUS_MPP_S     (1u << 11)  // MPP=01 (S-mode) at bits [12:11]
 
 // ── MMIO console (UPC L4f convention) ─────────────────────────────────────
+//
+// 2026-05-11: switched from `volatile uint32 *` to `volatile uint8 *`. The
+// previous u32 store at unaligned address 0xC001 was being lowered by GCC
+// into 4 separate byte stores (sb at 0xC001, 0xC002, 0xC003, 0xC004), of
+// which the first 3 hit the eBPF MMIO TTY intercept range (0xC000-0xC003)
+// and produced 3 bytes per call (char + 2 zero pad). The fourth store
+// landed in RAM at 0xC004 — wasted I/O.
+//
+// A direct byte store skips the splitting: one sb at 0xC001 -> one TTY
+// intercept -> one byte emitted. Drops the "x..v..6.. ..b.." pattern to
+// the clean "xv6 booting...\n" the operator actually wants to see.
 #define UPC_TTY_DATA_ADDR    0xC001
-#define UPC_TTY_REG          (*(volatile uint32 *)(UPC_TTY_DATA_ADDR))
+#define UPC_TTY_REG          (*(volatile uint8 *)(UPC_TTY_DATA_ADDR))
 
 static inline void mmio_putc(char c) {
-    UPC_TTY_REG = (uint32)c;
+    UPC_TTY_REG = (uint8)c;
 }
 
 static void mmio_puts(const char *s) {
